sap_integration_service/requirements.txt
Django>=5.0,<6
djangorestframework>=3.15,<4
psycopg[binary]>=3.1,<4
PyJWT[crypto]>=2.8,<3
cryptography>=42.0,<45
requests>=2.31,<3
celery[redis]>=5.3,<6
redis>=5.0,<6
python-dotenv>=1.0,<2
gunicorn>=22.0,<24
